show buttons on Log In
