+++ src/components/register.js
@@ -2,32 +2,28 @@ import { onNavigate } from '../main.js';
 import { registerUser } from '../lib/auth.js';
 
 export const register = () => {
-
   function registerUsuario() {
-    let valueEmail = document.getElementById('inputUsername').value;
-    let password = document.getElementById('inputPassword').value;
-  registerUser(valueEmail, password).then((userCredential) => {
+    const valueEmail = document.getElementById('inputUsername').value;
+    const password = document.getElementById('inputPassword').value;
+    registerUser(valueEmail, password).then((userCredential) => {
     // Signed in
-    const user = userCredential.user;
-    alert("usuario registrado correctamente")
-    console.log(user);
+      const user = userCredential.user;
+      alert('usuario registrado correctamente');
+      console.log(user);
     // ...
-  })
-  .catch((error) => {
-    const errorCode = error.code;
-    const errorMessage = error.message;
-    // ..
-    alert('nel carnal no te pueder registrar');
-  });
-};
+    })
+      .catch((error) => {
+        const errorCode = error.code;
+        const errorMessage = error.message;
+        // ..
+        alert('nel carnal no te pueder registrar');
+      });
+  }
 
   const div = document.createElement('div');
 
   const logoInicio = document.createElement('img');
   logoInicio.id = 'logo';
-
-  const logoGoogle = document.createElement('img');
-  logoGoogle.id = 'logoGoogle';
 
   const buttonLoginOne = document.createElement('button');
   buttonLoginOne.id = 'logIn2';
@@ -40,21 +36,13 @@ export const register = () => {
 
   const inputPassword = document.createElement('input');
   inputPassword.id = 'inputPassword';
-
-  const logoTwitter = document.createElement('img');
-  logoTwitter.id = 'logoTwitter';
-
-  const logogithub = document.createElement('img');
-  logogithub.id = 'logogithub';
+  inputPassword.setAttribute('type', 'password');
 
   logoInicio.src = './images/logo5.png';
   buttonLoginOne.textContent = 'Create acount';
   buttonBack.textContent = 'Back';
   inputEmail.placeholder = 'Email';
   inputPassword.placeholder = 'Password';
-  logoGoogle.src = './images/btn_google.png';
-  logoTwitter.src = './images/btn_twitter.png';
-  logogithub.src = './images/git.png';
 
   buttonLoginOne.addEventListener('click', () => {
     registerUsuario();
@@ -63,7 +51,7 @@ export const register = () => {
     onNavigate('/');
   });
 
-div.append(logoInicio, inputEmail, inputPassword, buttonLoginOne, buttonBack, logoGoogle, logoTwitter, logogithub);
+  div.append(logoInicio, inputEmail, inputPassword, buttonLoginOne, buttonBack);
 
   return div;
 };
