+++ src/lib/auth.js
@@ -1,7 +1,6 @@
-import { getAuth, createUserWithEmailAndPassword } from 'https://www.gstatic.com/firebasejs/9.9.3/firebase-auth.js'
-import {app} from './config.js'
+import { getAuth, createUserWithEmailAndPassword } from 'https://www.gstatic.com/firebasejs/9.9.3/firebase-auth.js';
+import { app } from './config.js';
 
 const auth = getAuth();
 
-export const registerUser =(email,password) => createUserWithEmailAndPassword(auth, email, password)
-
+export const registerUser = (email, password) => createUserWithEmailAndPassword(auth, email, password);
